requirements.txt
pygal
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==2.0.6
